requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
cryptography==42.0.5
Django==5.0.3
django-channels==0.7.0
django-filter==24.2
djangorestframework==3.15.1
idna==3.7
oauthlib==3.2.2
paypalrestsdk==1.13.3
pillow==10.3.0
pycparser==2.22
pyOpenSSL==24.1.0
python-decouple==3.8
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.11.0
urllib3==2.2.1
